based on async-mongo-2.0 starting harness v0.2.0-SNAPSHOT
