+++ java-sdk/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.actionml</groupId>
     <artifactId>harness-java-sdk</artifactId>
-    <version>0.1.1-RC4</version>
+    <version>0.2.0-SNAPSHOT</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
