+++ rest-server/admin/src/main/scala/com/actionml/admin/Administrator.scala
@@ -19,7 +19,7 @@ package com.actionml.admin
 
 import cats.data.Validated
 import cats.data.Validated.Invalid
-import com.actionml.core.template.Engine
+import com.actionml.core.engine.Engine
 import com.actionml.core.validate.{WrongParams, ValidateError}
 import com.typesafe.config.{Config, ConfigFactory}
 import com.typesafe.scalalogging.LazyLogging
