+++ rest-server/data/hinting.json
@@ -1,6 +1,6 @@
 {
   "engineId": "hinting",
-  "engineFactory": "com.actionml.templates.navhinting.NavHintingEngine",
+  "engineFactory": "com.actionml.engines.navhinting.NavHintingEngine",
   "algorithm":{
     "numQueueEvents": 50,
     "decayFunction": "click-order",
