+++ rest-server/data/test_resource.json
@@ -1,6 +1,6 @@
 {
   "engineId": "test_resource",
-  "engineFactory": "com.actionml.templates.cb.CBEngine",
+  "engineFactory": "com.actionml.engines.cb.CBEngine",
   "comment": "add mirrorContainer to point to a location on your system and enable mirroring to that location",
   "comment": "add sharedDBName for sharing a DBs for Users, should be named with lowercase and underscore separator, aka snake case",
   "sharedDBName": "harness_shared_user",
