+++ rest-server/data/test_resource_2.json
@@ -1,6 +1,6 @@
 {
   "engineId": "test_resource_2",
-  "engineFactory": "com.actionml.templates.cb.CBEngine",
+  "engineFactory": "com.actionml.engines.cb.CBEngine",
   "comment": "add sharedDBName for sharing a DBs for Users, should be named with lowercase and underscore separator, aka snake case",
   "algorithm":{
     "maxIter": 100,
