+++ rest-server/server/src/main/scala/com/actionml/router/service/EventService.scala
@@ -2,7 +2,7 @@ package com.actionml.router.service
 
 import cats.data.Validated.Invalid
 import com.actionml.admin.Administrator
-import com.actionml.core.template.Engine
+import com.actionml.core.engine.Engine
 import com.actionml.core.validate.{NotImplemented, WrongParams}
 import com.actionml.router.ActorInjectable
 import io.circe.syntax._
