--- rest-server/templates/src/main/resources/reference.conf
@@ -1,6 +1,0 @@
-mongo {
-  host = "localhost"
-  host = ${?MONGO_HOST}
-  port = 27017
-  port = ${?MONGO_PORT}
-}
